+++ src/RoadCaptain.App.Shared/Controls/SkiaPaints.cs
@@ -65,8 +65,11 @@ namespace RoadCaptain.App.Shared.Controls
         public static readonly SKPaint ElevationPlotPaint = new()
             { Color = SKColor.Parse("#000000"), Style = SKPaintStyle.Stroke, StrokeWidth = 2, IsAntialias = true };
         
-        public static readonly SKPaint ElevationPlotBackgroundPaint = new()
-            { Color = new SKColor(204, 204, 204, 100), Style = SKPaintStyle.Fill, IsAntialias = true };
+        public static readonly SKPaint ElevationPlotGradeZeroPaint = new SKPaint { Color = SKColor.Parse("#DD328BFF"), Style = SKPaintStyle.StrokeAndFill, StrokeWidth = 2, IsAntialias = true }; // blue
+        public static readonly SKPaint ElevationPlotGradeThreePaint = new SKPaint { Color = SKColor.Parse("#DD59BF59"), Style = SKPaintStyle.StrokeAndFill, StrokeWidth = 2, IsAntialias = true }; // green
+        public static readonly SKPaint ElevationPlotGradeFivePaint = new SKPaint { Color = SKColor.Parse("#DDFFCC3F"), Style = SKPaintStyle.StrokeAndFill, StrokeWidth = 2, IsAntialias = true }; // yellow
+        public static readonly SKPaint ElevationPlotGradeEightPaint = new SKPaint { Color = SKColor.Parse("#DDFF663A"), Style = SKPaintStyle.StrokeAndFill, StrokeWidth = 2, IsAntialias = true }; // orange
+        public static readonly SKPaint ElevationPlotGradeTenPaint = new SKPaint { Color = SKColor.Parse("#DDFF340C"), Style = SKPaintStyle.StrokeAndFill, StrokeWidth = 2, IsAntialias = true }; // red
 
         public static readonly SKPaint ElevationLinePaint = new()
             { Color = SKColor.Parse("#0000ff").WithAlpha((byte)(0xFF * 0.75)), Style = SKPaintStyle.Stroke, StrokeWidth = 1, PathEffect = SKPathEffect.CreateDash(new [] { 4f, 4f}, 1),IsAntialias = true };