+++ Changelog.md
@@ -13,6 +13,7 @@ Ultimately you _should_ notice less quirky behaviour at times which means you ca
 - Show sprints, KOMs and Strava segments along the route on the main screen [#51](https://github.com/sandermvanvliet/RoadCaptain/issues/51)
 ![Screenshot of Road Captain route builder showing special segments](./images/routebuilder-markers.png)
 - Spawn point now show the direction so it's easier to understand in which direction you can build a route [#48](https://github.com/sandermvanvliet/RoadCaptain/issues/48), [#49](https://github.com/sandermvanvliet/RoadCaptain/issues/49): ![Direction marker on spawn point segment](./images/spawn-point-direction.png)
+- The elevation plot now has colorized grade indications so you can see more easily how hard the cilmbs are! ![Color grading on elevation plot](./images/route-builder-grades.png)
 
 ### Runner